Split the carrier harness into per-topic submodules
tests/carrier_generic.rs had grown to one ~1000-line file. Break the generic
bodies out at natural boundaries into carrier_generic/: arithmetic, shifts and
bitwise, bit inspection + comparison, and numeric. The shared Carrier surface,
the for_both_carriers! runner, and the width-pinning from_u32 constructor move
to a harness submodule the others pull in.

The bodies are unchanged — only relocated — so coverage is identical. The root
file is now module wiring; submodules reach the harness via `use
crate::harness::*`. `#[path]` points at the subdirectory because an
integration-test root would otherwise resolve `mod foo` to a sibling
`tests/foo.rs` and compile it as a separate test binary.
